🧹 [code health] refactor AdminDashboard components and logic into smaller modular files
